started to shift entity attributes to variables
Former-commit-id: aa71905e100fa7b73bacc4716aa571ef8e6f6895

+++ src/catalogue/DatasetProfile.py
@@ -9,7 +9,7 @@ class DatasetProfile:
     description: str
     theme: str
     filetype: str
-    size: int
+    size: float  # in Mb
     nb_tuples: int
     completeness: int
     uniqueness: float
